apply bridge pattern
