+++ app/services/location/sourcelocations.py
@@ -1,0 +1,8 @@
+from abc import *
+
+class DataSourcesInterface(ABC):
+    @abstractmethod
+    async def get_locations(self):
+        pass
+
+
